server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
Pillow
PyPDF2
pdf2image
pydantic
pydantic-settings

